app.js code restructure
